cleanup

+++ csrc/ir_internal_nodes.h
@@ -58,7 +58,7 @@ class TORCH_CUDA_CU_API SelectOp : public Expr {
  public:
   using Expr::Expr;
 
-  SelectOp(IrBuilderPasskey, Val* out, Val* in, int dim, Val* index);
+  SelectOp(IrBuilderPasskey, Val* out, Val* in, int64_t dim, Val* index);
 
   NVFUSER_DECLARE_CLONE_AND_CREATE
 
@@ -69,8 +69,8 @@ class TORCH_CUDA_CU_API SelectOp : public Expr {
   std::string toString(int indent_size = 0) const override;
   std::string toInlineString(int indent_size = 0) const override;
 
-  int dim() const {
-    return attribute(0)->as<Attribute<int>>()->value;
+  int64_t dim() const {
+    return attribute(0)->as<Attribute<int64_t>>()->value;
   }
 
   IterDomain* getIndexedID() const;
@@ -84,7 +84,7 @@ class TORCH_CUDA_CU_API IndexSelectOp : public Expr {
  public:
   using Expr::Expr;
 
-  IndexSelectOp(IrBuilderPasskey, Val* out, Val* in1, int dim, Val* in3);
+  IndexSelectOp(IrBuilderPasskey, Val* out, Val* in1, int64_t dim, Val* in3);
 
   NVFUSER_DECLARE_CLONE_AND_CREATE
 
@@ -107,8 +107,8 @@ class TORCH_CUDA_CU_API IndexSelectOp : public Expr {
 
   IterDomain* getConsumerOfIndexedID() const;
 
-  int dim() const {
-    return attribute(0)->as<Attribute<int>>()->value;
+  int64_t dim() const {
+    return attribute(0)->as<Attribute<int64_t>>()->value;
   }
 };
 
@@ -125,7 +125,7 @@ class TORCH_CUDA_CU_API TorchGatherOp : public Expr {
       IrBuilderPasskey,
       Val* out,
       Val* in,
-      int dim,
+      int64_t dim,
       Val* index,
       bool exact_sizes);
 
@@ -146,8 +146,8 @@ class TORCH_CUDA_CU_API TorchGatherOp : public Expr {
     return input(1)->as<TensorView>();
   }
 
-  int dim() const {
-    return attribute(0)->as<Attribute<int>>()->value;
+  int64_t dim() const {
+    return attribute(0)->as<Attribute<int64_t>>()->value;
   }
 
   IterDomain* getIndexedID() const;
@@ -167,7 +167,7 @@ class TORCH_CUDA_CU_API ScatterOp : public Expr {
       ScatterOpType type,
       Val* out,
       Val* self,
-      int dim,
+      int64_t dim,
       Val* index,
       Val* src);
 
@@ -192,8 +192,8 @@ class TORCH_CUDA_CU_API ScatterOp : public Expr {
     return input(2)->as<TensorView>();
   }
 
-  int dim() const {
-    return attribute(0)->as<Attribute<int>>()->value;
+  int64_t dim() const {
+    return attribute(0)->as<Attribute<int64_t>>()->value;
   }
 
   IterDomain* getIndexedID() const;
